backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.18
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-docx==1.1.2
pypdf==4.3.1
aiofiles==24.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.4.0
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
-e .
